feat: add newsboat filetype
Adds a filetype definition for `newsboat` files, associating the
`.newsboat` extension with the `newsboat` filetype. This allows
Neovim to apply appropriate syntax highlighting and other filetype-
specific settings to these files.

feat: add citation picker with bibtex support

This commit introduces a new citation picker feature that allows
users to easily insert citation keys from a BibTeX file into their
documents.

The citation picker uses `bibtex2csv` to parse the BibTeX file and
displays the entries in a fuzzy-searchable list. Users can select
one or more entries, and the corresponding citation keys will be
inserted into the current buffer in Pandoc format.

The commit also includes the following changes:

- Adds a new keymap `<leader>ic` to open the citation picker.
- Adds a new keymap `<localleader>ic` to open the citation picker
  in insert mode.
- Adds a new `citation` function to the `pickers` module.
- Adds a new `pick` function to the `pickers` module.
- Removes the `pick_cmd_result` function from the `pickers` module.
- Updates the `fasd` function to use the new `pick` function.
- Disables the old citation keymaps in `fzf-lua.lua`.
- Adds a new `picker` configuration option to the `snacks` plugin.

The citation picker is designed to be used with the
`bibtex-cite` command-line tool, which is not included in this
commit. Users will need to install `bibtex-cite` separately.

The citation picker also supports previewing the file associated
with a citation key. This requires the `mylib` command-line tool,
which is also not included in this commit. Users will need to
install `mylib` separately.

The citation picker also supports bracket citations. This can be
triggered by pressing `<c-i>` in the picker window.

refactor: remove unused code and improve picker API

This commit removes the `pick_cmd_result` function from the
`pickers` module, as it is no longer used. It also introduces a
new `pick` function that provides a more flexible and consistent
API for creating pickers.

The `fasd` function has been updated to use the new `pick`
function.

The `snacks` plugin has been updated to include a new `picker`
configuration option, which allows users to customize the behavior
of the picker window.

The `fzf-lua.lua` plugin has been updated to disable the old
citation keymaps, as they are no longer needed.