feat(lazyvim.json): add yanky, inc-rename, mini-hipatterns feat(lazyvim.json): add snacks_explorer feat(after/ftplugin/quarto.lua): remove syntax match for QuartoCrossRef
The syntax match was removed because it was causing
issues with the highlighting of other elements in the
document.

refactor(plugins/editor.lua): refactor yanky.nvim config

The yanky.nvim config was refactored to use the
snacks.nvim picker instead of the default yanky.nvim
picker. This was done to improve the user experience.

feat(plugins/ui.lua): add mini.hipatterns plugin

The mini.hipatterns plugin was added to provide
highlighting for citations in quarto, rmarkdown, and
markdown files.